EditNickNameForm added, PersonalInformation refactored
